Small refactoring to rename aoc2 input helper.
