+++ src/aoc2/aoc2.py
@@ -1,5 +1,5 @@
 import src.utils.file_helper as file_helper
-import src.aoc2.input_helper as input_helper
+import src.aoc2.aoc2_input_helper as input_helper
 from src.aoc2.bag import Bag
 from src.aoc2.elf_game import ElfGame
